refactor(core,additional-components): move panel component to core pkg
