--- packages/additional-components/src/panel/index.ts
@@ -1,2 +1,0 @@
-export { default as Panel } from './Panel.vue'
-export * from './types'
