+++ packages/core/src/types/index.ts
@@ -8,3 +8,4 @@ export * from './store'
 export * from './hooks'
 export * from './changes'
 export * from './handle'
+export * from './panel'
