+++ packages/core/src/types/panel.ts
